# CTR-DRBG: Use hardware acceleration on X86